Make Original StackView to share.
