+++ Queens-game/Content/CitizenSelected/CitizenSelectedViewController.swift
@@ -11,9 +11,13 @@ import RxCocoa
 
 class CitizenSelectedViewController: UIViewController {
 
+  let disposeBag = DisposeBag()
+
   let viewModel = CitizenSelectedViewModel()
 
-  let disposeBag = DisposeBag()
+  lazy var countdownStackView = CountdownStackView(
+    countdown: self.viewModel.countdonwTime
+  )
 
   let screenTitle: H2Label = {
     let lb = H2Label(text: "Selecting the citizen...")
@@ -31,83 +35,11 @@ class CitizenSelectedViewController: UIViewController {
     return button
   }()
 
-  let suits: [UILabel] = {
-    let suits = ["♠","♦","♥","♣"]
-    let suitLabels: [UILabel] = suits.map { suit in
-      let label = UILabel()
-      label.text = suit
-      label.font = CustomFont.h2
-      if ["♦","♥"].contains(suit) {
-        label.textColor = .red
-      }
-      return label
-    }
-    return suitLabels
-  }()
-
-  lazy var above: HorizontalStackView = {
-    let stackView = HorizontalStackView(
-      arrangedSubviews: [suits[0], suits[1]],
-      distribution: .equalSpacing
-    )
-    return stackView
-  }()
-
-  lazy var countdownLabel: UILabel = {
-    let label = UILabel()
-    label.text = String(self.viewModel.countdonwTime)
-    label.font = CustomFont.h1
-    label.textAlignment = .center
-    label.sizeToFit()
-    return label
-  }()
-
-  lazy var sec: UILabel = {
-    let label = UILabel()
-    label.text = "sec"
-    label.textAlignment = .center
-    return label
-  }()
-
-  lazy var center: VerticalStackView = {
-    let stackView = VerticalStackView(
-      arrangedSubviews: [countdownLabel, sec],
-      distribution: .equalSpacing
-    )
-    return stackView
-  }()
-
-  lazy var below: HorizontalStackView = {
-    let stackView = HorizontalStackView(
-      arrangedSubviews: [suits[2], suits[3]],
-      distribution: .equalSpacing
-    )
-    return stackView
-  }()
-
-  lazy var countdownBlock: VerticalStackView = {
-    let stackView = VerticalStackView(
-      arrangedSubviews: [
-        self.above,
-        self.center,
-        self.below
-      ]
-    )
-    stackView.spacing = 32
-    return stackView
-  }()
-
-  lazy var wrapperCountdownBlock: VerticalStackView = {
-    let stackView = VerticalStackView(arrangedSubviews: [countdownBlock])
-    stackView.alignment = .center
-    return stackView
-  }()
-
   lazy var stackView: VerticalStackView = {
     let sv = VerticalStackView(
       arrangedSubviews: [
         screenTitle,
-        wrapperCountdownBlock,
+        countdownStackView,
         skipButton
       ]
     )
@@ -119,10 +51,10 @@ class CitizenSelectedViewController: UIViewController {
   override func viewDidLoad() {
     super.viewDidLoad()
     setupLayout()
-    //    self.viewModel.countdown()
+//    self.viewModel.countdown()
     self.viewModel.rxCountdownTime
       .subscribe(onNext: { [weak self] time in
-        self?.countdownLabel.text = String(time!)
+        self?.countdownStackView.countdownLabel.text = String(time!)
       },
       onCompleted: {
         self.replaceView()
@@ -155,14 +87,10 @@ extension CitizenSelectedViewController {
       equalTo: view.safeAreaLayoutGuide.trailingAnchor,
       constant:  Constant.Common.trailingSpacing
     ).isActive = true
-    countdownBlock.widthAnchor.constraint(
-      equalTo: wrapperCountdownBlock.widthAnchor,
-      multiplier: 0.8
-    ).isActive = true
   }
 
   @objc func skipButtonTapped(_ sender: UIButton) {
-    self.stackView.removeFromSuperview()
+    self.viewModel.rxCountdownTime.onCompleted()
     self.toResult()
   }
